Oracleのとき、boolean型をサポートするよう修正
